refactor: Reorganize directory structure for better organization
- Consolidates utility and config files into lib/ directory
- Groups resource-related files under app/(resources)/
- Simplifies root page location

- List of details:
  - Move metadata config from config/ to lib/
  - Relocate global.css from config/ to app/(resources)/
  - Move all utility files from utils/ to lib/
  - Relocate main page from app/(index)/ to app/
  - Move feed-related files from app/(files)/ to app/(resources)/(feed)/
