+++ src/app/(resources)/globals.css
