+++ src/components/layouts/Footer.tsx
@@ -1,12 +1,7 @@
 'use client'
 
 import ExternalLink from '@/components/ui/ExternalLink'
-import {
-  ownerName,
-  ownerGithub,
-  ownerEmail,
-  ownerDiscord,
-} from '@/config/metadata'
+import {ownerName, ownerGithub, ownerEmail, ownerDiscord} from '@/lib/metadata'
 import {Github, Mail, MessageCircle, Rss} from 'lucide-react'
 
 export default function Footer() {
